Tightening a few more printf-type arguments.

+++ output/latex/message.c
@@ -16,7 +16,7 @@
 void output_latex_message(FILE *file, message_id_t id, ...)
 {
   va_list args;
-  DBG((stderr, "Mesage(%d) = %s\n", id, output_message_get(id)));
+  DBG((stderr, "Mesage(%d) = %s\n", (int) id, output_message_get(id)));
   va_start(args,id);
   if (id<MsgCount)
     vfprintf(file,output_message_get(id),args);
